+++ lib/fix-ical-content.js
