+++ lib/get-ical-content.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const ical = require('node-ical');
-const icalFix = require('./icalFix');
+const fixContent = require('./fix-ical-content');
 
 module.exports = (url) => {
     return new Promise((resolve, reject) => {
@@ -14,7 +14,7 @@ module.exports = (url) => {
                     reject(data);
                 }
                 else {
-                    resolve(icalFix(data.body));
+                    resolve(fixContent(data.body));
                 }
             }
         });
